# specs/001-matrades-product-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Matrades Application API
  version: 1.1.0
  description: >-
    Contract for authenticated configuration, research, strategy, risk, approval,
    reconciliation, monitoring, journal, and performance workflows. Broker execution
    writes are intentionally absent from V1.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /instruments:
    get:
      tags: [Research]
      operationId: listInstruments
      parameters:
        - name: asset_class
          in: query
          schema: { $ref: '#/components/schemas/AssetClass' }
        - name: instrument_type
          in: query
          schema: { $ref: '#/components/schemas/InstrumentType' }
        - name: connection_id
          in: query
          schema: { type: string, format: uuid }
        - name: capability
          in: query
          schema: { $ref: '#/components/schemas/MarketDataCapability' }
      responses:
        '200':
          description: Canonical instruments and exact provider or broker listings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/InstrumentDescriptor' }
        default: { $ref: '#/components/responses/Error' }
...
  /trade-proposals/{proposal_id}:
    get:
      tags: [Trading]
      operationId: getTradeProposal
      parameters:
        - $ref: '#/components/parameters/ProposalId'
      responses:
        '200':
          description: Proposal with complete equity snapshot and validation evidence
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TradeProposal' }
        default: { $ref: '#/components/responses/Error' }
  /trade-proposals/{proposal_id}/decisions:
    post:
      tags: [Trading]
      operationId: decideTradeProposal
      parameters:
        - $ref: '#/components/parameters/ProposalId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Hil2Decision' }
      responses:
        '200':
          description: TAKE, WAIT, or REJECT recorded; no broker write
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApprovalDecision' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410':
          description: Proposal expired or requires revalidation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        default: { $ref: '#/components/responses/Error' }
...
  /events:
    get:
      tags: [Operations]
      operationId: streamOperationalEvents
      parameters:
        - name: Last-Event-ID
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Owner-scoped SSE projection; clients re-fetch after version gaps
          content:
            text/event-stream:
              schema: { type: string }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: matrades_session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Expected aggregate/configuration version
      schema: { type: string }
...
    ProposalId:
      name: proposal_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Structured safe error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Idempotency, state-transition, or optimistic-concurrency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, correlation_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        correlation_id: { type: string, format: uuid }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: string, pattern: '^-?[0-9]+(\.[0-9]+)?$' }
        currency: { type: string, minLength: 3, maxLength: 3 }
    AccountSnapshot:
      type: object
      required: [id, account_id, starting_balance, balance, equity, floating_pnl,
        realized_daily_pnl, current_drawdown, daily_drawdown, source_time, freshness]
      properties:
        id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        starting_balance: { $ref: '#/components/schemas/Money' }
        balance: { $ref: '#/components/schemas/Money' }
        equity: { $ref: '#/components/schemas/Money' }
        floating_pnl: { $ref: '#/components/schemas/Money' }
        realized_daily_pnl: { $ref: '#/components/schemas/Money' }
        current_drawdown: { $ref: '#/components/schemas/Money' }
        daily_drawdown: { $ref: '#/components/schemas/Money' }
        source_time: { type: string, format: date-time }
        freshness: { enum: [VALID, SUSPECT, STALE, INVALID] }
    EffectiveConstraint:
      type: object
      required: [type, effective_value, controlling_source, severity]
      properties:
        type: { type: string }
        effective_value: { type: object, additionalProperties: true }
        controlling_source: { type: string }
        severity: { enum: [INFO, WARNING, SOFT_BLOCK, HARD_BLOCK] }
        contributors:
          type: array
          items: { type: string }
...
    RiskCapacityResult:
      type: object
      required: [id, status, existing_reserved_risk, candidate_risk,
        projected_portfolio_risk, remaining_daily_capacity, remaining_total_capacity,
        additional_trade_capacity, limiting_constraints, expires_at]
      properties:
        id: { type: string, format: uuid }
        status: { enum: [PASS, REDUCE_SIZE, HARD_BLOCK] }
        compliant_size: { type: string }
        existing_reserved_risk: { $ref: '#/components/schemas/Money' }
        candidate_risk: { $ref: '#/components/schemas/Money' }
        projected_portfolio_risk: { $ref: '#/components/schemas/Money' }
        remaining_daily_capacity: { $ref: '#/components/schemas/Money' }
        remaining_total_capacity: { $ref: '#/components/schemas/Money' }
        additional_trade_capacity: { type: integer, minimum: 0 }
        limiting_constraints:
          type: array
          items: { $ref: '#/components/schemas/EffectiveConstraint' }
        expires_at: { type: string, format: date-time }
    AssetClass:
      type: string
      enum: [FOREX, METALS, CRYPTOCURRENCY, STOCKS]
    InstrumentType:
      type: string
      enum: [SPOT, CFD, FUTURES]
    MarketDataCapability:
      type: string
      enum: [INSTRUMENT_DIRECTORY, DISCOVERY, QUOTE, TRADES, CANDLES, ORDER_BOOK,
        CONTRACT_DETAILS, FUTURES_CHAIN, OPEN_INTEREST, FUNDING, CORPORATE_ACTIONS,
        BROKER_TRADABILITY, ACCOUNT_BALANCES]
...
    InstrumentSpecification:
      type: object
      required: [id, version, effective_from, price_currency, quantity_unit, tick_size,
        quantity_minimum, quantity_step, provenance, freshness]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        effective_from: { type: string, format: date-time }
        effective_to: { type: [string, 'null'], format: date-time }
        price_currency: { type: string }
        settlement_currency: { type: string }
        quantity_unit: { enum: [UNITS, SHARES, LOTS, CONTRACTS] }
        contract_multiplier: { type: string }
        tick_size: { type: string }
        tick_value: { type: string }
        quantity_minimum: { type: string }
        quantity_maximum: { type: string }
        quantity_step: { type: string }
        margin_terms: { type: object, additionalProperties: true }
        financing_terms: { type: object, additionalProperties: true }
        expiry: { type: [string, 'null'], format: date-time }
        first_notice: { type: [string, 'null'], format: date-time }
        last_trade: { type: [string, 'null'], format: date-time }
        provenance: { type: object, additionalProperties: true }
        freshness: { enum: [VALID, SUSPECT, STALE, INVALID] }
    InstrumentDescriptor:
      type: object
      required: [instrument_id, venue_instrument_id, venue, symbol, asset_class,
        instrument_type, specification]
      properties:
        instrument_id: { type: string, format: uuid }
        venue_instrument_id: { type: string, format: uuid }
        venue: { type: string }
        symbol: { type: string }
        asset_class: { $ref: '#/components/schemas/AssetClass' }
        instrument_type: { $ref: '#/components/schemas/InstrumentType' }
        futures_contract_id: { type: [string, 'null'], format: uuid }
        continuous_series: { type: boolean, default: false }
        executable: { type: boolean, default: true }
        specification: { $ref: '#/components/schemas/InstrumentSpecification' }
...
    TradeConstruction:
      type: object
      required: [instrument_id, venue_instrument_id, specification_version_id, asset_class,
        instrument_type, direction, entry, stop_loss, targets, requested_size, quantity_unit]
      properties:
        instrument_id: { type: string, format: uuid }
        venue_instrument_id: { type: string, format: uuid }
        specification_version_id: { type: string, format: uuid }
        futures_contract_id: { type: [string, 'null'], format: uuid }
        asset_class: { $ref: '#/components/schemas/AssetClass' }
        instrument_type: { $ref: '#/components/schemas/InstrumentType' }
        direction: { enum: [BUY, SELL] }
        entry: { type: string }
        entry_zone: { type: object, additionalProperties: true }
        stop_loss: { type: string }
        targets: { type: array, minItems: 1, items: { type: string } }
        requested_size: { type: string }
        quantity_unit: { enum: [UNITS, SHARES, LOTS, CONTRACTS] }
        invalidation: { type: object, additionalProperties: true }
    TradeProposal:
      type: object
      required: [id, account_id, state, construction, strategy_version_id,
        market_fingerprint_id, account_snapshot, risk_result, policy_status,
        guardrail_status, critic_status, expires_at, version]
      properties:
        id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        state: { enum: [TRADE_PENDING_APPROVAL, AWAITING_MANUAL_ENTRY, WAITING,
          REJECTED, EXPIRED, REVALIDATION_REQUIRED] }
        construction: { $ref: '#/components/schemas/TradeConstruction' }
        strategy_version_id: { type: string, format: uuid }
        market_fingerprint_id: { type: string, format: uuid }
        evidence_refs: { type: array, items: { type: string } }
        account_snapshot: { $ref: '#/components/schemas/AccountSnapshot' }
        risk_result: { $ref: '#/components/schemas/RiskCapacityResult' }
        policy_status: { enum: [PASS, WARNING] }
        guardrail_status: { enum: [PASS, WARNING] }
        critic_status: { enum: [PASS] }
        expires_at: { type: string, format: date-time }
        version: { type: integer }
    Hil2Decision:
      type: object
      required: [action]
      properties:
        action: { enum: [TAKE, WAIT, REJECT] }
        rationale: { type: string, maxLength: 2000 }
...
    ApprovalDecision:
      type: object
      required: [id, approval_type, action, resolved_at, resulting_subject_version]
      properties:
        id: { type: string, format: uuid }
        approval_type: { enum: [MARKET_SELECTION, TRADE_ENTRY, TRADE_MANAGEMENT] }
        action: { type: string }
        resolved_at: { type: string, format: date-time }
        resulting_subject_version: { type: integer }
...